Improve design of "See all" link in feat. section
This commit improves the design of the "See all" link in the featured
versions section. Previously it used to be a link in parentheses next to
the heading, which did not look great and didn't semantically make
sense. Now the link is separated, aligned to the right and has chevron
to make it look like a clickable object. The clickable area has been
increased too, and is displayed on hover.