i18n-formatted component and i18n directive
Adds i18n-formatted functional component which accepts properties like
message-id (identifier for the message) and values (object with custom
values). It then formats the string under the provided message-id, using
provided values either through arguments or as children, for which it
uses 0-based number of occurance or custom name set by the new
v-i18n:value directive.

Say if we have the following in our messages:

  {
    "my-message": "Hello, {0}!"
  }

Then we can use i18n-formatted as such:

  <i18n-formatted message-id="my-message">
    <strong>{{ greetingTarget }}</strong>
  </i18n-formatted>

(Assuming greetingTarget is coming from the parent component)

This should result into the following HTML:

  Hello, <strong>World</strong>!

Of course, that '0' in our message doesn't speak value for translators,
so we better give it a good name like 'target'. For this we use the
v-i18n directive with value argument:

  <i18n-formatted message-id="my-message">
    <strong v-i18n:value="'target'">{{ greetingTarget }}</strong>
  </i18n-formatted>

(Don't forget that all directives accept expressions by default!)

And then, of course, change our message:

  {
    "my-message": "Hello, {target}!"
  }

This still would result in the same HTML, but now it's more clear which
value will be used how, the argument in the message has also became
meaningful for translators - they now can know the placeholder is likely
to be replaced by the user name or something else.