Add preferred language saving
Now, whenever user changes their preferred language, it will be saved to
both cookies and localStorage for persistence. They will be restored the
next time user opens the site.

The option to sync language with the browser language also was added, it
will be used by default if user newer changed the languge or if they
manually selected it.

When this option is selected the site will use closest possible language
that is available. If there is no closest language, the default language
will be selected.

If user changes their language while browsing the site, and they have
'Sync with the browser' option selected, the site will automatically
change the language, no refresh needed.