Add v-i18n:wrap directive
v-i18n:wrap directive is a directive that only has effect within the
<i18n-formatted> element. It takes an element as a skeleton and allows
to use so-called 'wrappers' in the translations.

Wrappers are extremely useful if translation needs to contain links or
buttons, but it does not make to split text from those to separate
strings.

For example, let's say we have the following HTML:

  Your project was rejected, please check out the <a href="/rules">rules</a>.

Without wrapper elements we would have to resort to the following:

In our translations file,

  {
    "rejection-message": "Your project was rejected, please check out the {rulesLink}.",
    "rejection-message.rules-link": "rules"
  }

In our Vue file,

  <i18n-formatted message-id="rejection-message">
    <a href="/rules" v-i18n:value="'rules-link'">{{ $t('rejection-message.rules-link') }}</a>
  </i18n-formatted>

For translators, this is not very plausable experience because they'd
have to guess what that 'rulesLink' variable actually contains, it's
good that we use good message IDs, but nevertheless, it's confusing
experience.

It also introduces a surface for mistake, translator may not pay close
attention to the message IDs and translate both strings independently.
As English speakers we tend to forget that other languages have more
advanced casing systems, where noun or verb may change depending on what
surrounds it, so in some languages, translating both strings can lead to
mistake.

To demonstrate, let me give you an example from Discord:

Say we have the following strings in original English text:

  {
    "hello-cta": "Say ‘{hello}’ to the {helloTarget}",
    "hello-cta.target": "cat"
  }

Which we can use as such:

  $t('hello-cta', { helloTarget: $t('hello-cta.target') })

But what happens if Ukrainian translator translates `hello-cta.target`
independently of `hello-cta` string. They translate it as such:

  {
    "hello-cta": "Сказати «привіт» {helloTarget}",
    "hello-cta.target": "кіт"
  }

And using code above, the result we will get from these translations:

  'Сказати «привіт» кіт'

Which is incorrect, in this case 'кіт' is supposed be in dative case,
which would change it to 'коту'.

Keeping that in mind, let's look how things would change if we were to
use wrapping syntax:

  {
    "hello-cta": "Сказати «привіт» <target>коту</target>"
  }

This looks like HTML, and if you think about it, it kinda is, but not
quite, it's merely wrapper, like BB-codes, and those do not accept any
arguments, to not complicate things.

In the code we would use `hello-cta` string as such:

  $t('hello-cta', { helloTarget: (parts) => `${parts.map(_ => String(_)).join('')}` }`)

Keep in mind, that in this example we don't do anything fancy yet, you
can see that our wrapper doesn't do anything apart from taking some
parts and later converting them to strings. If we don't use any objects
in our context object, then it will never be not array of strings, but
otherwise it'd contain these objects too in places of variables.

So what do we get when evaluating code above? Well, of corse:

  'Сказати «привіт» коту'

What v-i18n:wrap allows us to do is to create skeleton node, which will
be copied every time the wrapper used, and which children will be
replaced with aforementioned parts.

Let's look at the first example, but using wrappers now:

Here's our changed translation file:

  {
    "rejection-message": "Your project was rejected, please check out the <rules-link>rules</rules-link>."
  }

And here's how we'd change our template:

  <i18n-formatted message-id="rejection-message">
    <a href="/rules" v-i18n:wrap="'rules-link'" />
  </i18n-formatted>

Notice that we self-close our link element, that's because we know that
anything will be replaced anyway, so there is no sense to provide any
children.

And what happens when this string actually gets rendered?

  Your project was rejected, please check out the <a href="/roles">rules</a>.

Exactly what we'd expect.

So there we go, this is the new directive for you, hope this was
insightful and useful to understand the importance and usage of this
feature.