Add formatter for compact numbers
Unfortunately, Intl.NumberFormat API has a serious weakness when it
comes to using the numbers in compact notation in translations: you
cannot choose correct plural form based on the result it provides.
That's because the result is a completely formatted string, which you
cannot reliably parse for all the locales.

Many consider that passing the original number for plural selection
will work, but that's not the case, because languages actually have
declensions for the words coming before or after the number in compact
notation. Not to mention, the very number for compact unit can decline
based on the number in compact notation, which is different from the
original number.

For example, here's how English cats translate to Ukrainian:

* 0.5 cats ⇒ 0,5 кота.
* 1 cat => 1 кіт.
* 2 cats => 2 коти.
* 5 cats => 5 котів.
* 1 thousand cats => 1 тисяча котів.
* 2 thousand cats => 2 тисячі котів.
* 5 thousand cats => 5 тисяч котів.

Pay close attention to how Ukrainian changes when talking about
thousands. Nothing magical happening here: it's just that instead of
cats, it starts to decline word for 'thousand' instead, and word for
'cats' goes into genitive plural case.

Now, what would happen if we were to use compact notation?

* 0.5 cats ⇒ 0,5 кота.
* 1 cat => 1 кіт.
* 2 cats => 2 коти.
* 5 cats => 5 котів.
* 1K cats => 1 тис. котів.
* 2K cats => 2 тис. котів.
* 5K cats => 5 тис. котів.

English just changed to use 'K' after the number, yet Ukrainian kept the
word for 'thousand' there, just shortened it ever so slightly. Yes, the
word is still here, so keep that in mind.

It's important that I gave you example for small numbers, including
decimals, because they demonstrate all four plural forms Ukrainian has.

Let's write that cat format using ICU Message Format:

  {cats, count, one {# cat} other {# cats}}

Yep. It's that simple in English because it uses just two plural forms,
1 (and -1, but that's controversial) for 'one', and anything else goes
to 'other'.

Now let's translate it to Ukrainian:

  {cats, count, one {# кіт} few {# коти} many {# котів} other {# кота}}

Nothing suspicious yet.

But then let's use this translation with the following numbers of cats:

* 1,500 ⇒ 1,5 тис. котів.
* 1,501 ⇒ 1,5 тис. кіт.
* 1,502 ⇒ 1,5 тис. коти.
* 1,505 ⇒ 1,5 тис. котів.

If you've read the above carefully, you already spotted the mistake.
This is the weakness of JavaScript Intl API. If you correctly use ICU
library for Java, it will, in fact, correctly apply the plural, because
when you convert number to compact notation in ICU it does not give you
the formatted string, it gives you a special object that you can then
use as context value to format your message. Intl.NumberFormat in
JavaScript, unfortunately, returns plain string, and even formatting to
parts can't help you.

So how do we go to fix this problem? The writers of the original
Intl.NumberFormat proposal already work on finding a way to mitigate
this issue, but it will take time, it already took enormous amount of
time while sites continue fall for this pitfall.

The solution is not easy: you need to use original CLDR data to
correctly calculate number on which you should determine the plural
form. Glad for us, there's a polyfill for Intl.NumberFormat with its
internals somewhat exposed. We can use — abuse, really — these internals
to perform what Intl does behind the scenes: calculate exponents and
actual number value.

Unfortunately, this comes with a significant drawback: you must
duplicate context values twice: once for formatted number (in compact
notation), and then for number that you actually use to select the
plural form. This can be somewhat confusing to translators, but it's
what we got to do to mitigate the obvious mistake.

I created a simple formatCompactNumber function within IntlShape, which
accepts the value to format and almost all — except for notation,
obviously — number formatting options; it then returns a tuple that
contains the formatted number and number for pluralisation.

Because it's a tuple, we'd want to give it a meaningful value, for what
I created another helper function — deunionize. It accepts a tuple, or
any array, really, and keys as parameters. It then creates hashmap (an
object without any prototype) for every key assigns value from array, in
order of passed keys.

Thus, we can do this when translating strings that use counters:

In translation file,

  {
    "downloads": "{downloads, plural, one {{downloadsCounter} download} other {{downloadsCounter} downloads}}"
  }

In code,

  $t('downloads', $deunionize($fmt.formatCompactNumber(downloads), 'downloadsCounter', 'downloads'))

deunionize function will create us the following hashmap (for 1,501):

  {
    downloadsCounter: "1,5K",
    downloads: 1_500,
  }

For more complex context values objects spread can be used.