Adopt ago() from Omorphia as $fmt helper
Omorphia, a design system for Modrinth Rewrite, has useful utility `ago`
used to display relative times in human readable form (e.g. 1 hour ago,
in 1 day, etc). It's going to be useful to have it as part of $fmt
helpers of our internationalization plugin.

It's somewhat in syntax, namely there are no different parameters for
`date` and `nowDate`, instead it uses date directly, or array of two
dates, `from` and `to`, which translate to what parameters `date` and
`nowDate` do in Omorphia helper.