Introduce IntlFormatted component to i18n plugin
IntlFormatted is a rewritten i18n-formatted component that uses slots
instead of the magical directive, providing more verbose, but cleaner
templating, without weird magic. It also gives ability to create nested
wrappers because children are now passed as slot bindings rather than
inserted directly into wrapper element.

Here's an example usage of the new component. It demonstrates most of
the IntlFormatted features, but you probably won't use it like that
(e.g. you'll merge target and ~targetName slots, etc).

  <IntlFormatted
    message="<i>Hello, <target>{targetName}</target>!</i>"
    :tags="['i']"
    :values="{ $name: 'Brawaru' }"
  >
    <template #target="{ children }">
      <strong><Fragment :of="children" /></strong>
    </template>
    <template #~targetName="{ values: { $name } }">
      {{ $name }}
    </template>
  </IntlFormatted>

There are a few things to unpack in this example:

1. We provide a custom message to the component, in that message we use
   both values (targetName) and wrappers (i, target).

2. We then define standard tags that will be provided automatically,
   namely 'i', an element for italic text.

3. We then provide values to format message with. Because we don't want
   $name value to be used in texts, as it is a computed property, we
   prefix its name with $, which would make it invalid key name within
   ICU Message Format syntax.

4. Then we define template for 'target' slot. Slot names can be
   anything, they will be used as a name for the wrapper. For wrapper
   slots we can expect the properties 'children' (formatted contents
   inside the wrapper) and 'values' (the initially provided values to
   format message with).

  (i) It is recommended to use values in the formatted message when
      possible instead of using the one from object. They are provided
      for the convenience of accessing computed values (such as compact
      numbers).

5. Finally, we also define template for '~targetName' slot. Notice '~'
   character in the slot name. All tilde slots are provided as a value
   for formatting the message, not as a wrapper. Therefore it will not
   have any 'children' in slot properties, only 'values'.

The above component will render to this HTML:

  <i>Hello, <strong>Brawaru</strong>!</i>

As you can see, the new component is quite powerful, verbose, but easier
to understand of what's going on. Previous component is now deprecated
and will be removed in the following commits, as soon as all of its
usage will be replaced with the new component.