Properly handle value slots in IntlFormatted comp.
All slots return arrays of elements, so values for immediately invoked
slots (aka value slots) would result in 'formatted' containing sub-array
of VNodes, which was converted to text since Array is not a VNode on its
own. To handle this just flattening the array one time seems to do the
job.