Add trim option to messageLoader
It is true by default and trims the text contents of the file before
parsing it as a message. This helps to avoid unexpected line breaks at
the end of messages since most files have empty line at the end. In
fact, those line breaks are so unexpected that they completely break
hydration due to minifier yeeting the line breaks. Lovely.